Modularize encrypted replica pipeline
